+++ Source/DlgSystem/Private/DlgManager.cpp
@@ -127,8 +127,7 @@ bool UDlgManager::CanStartDialogue(UDlgDialogue* Dialogue, UPARAM(ref)const TArr
 		return false;
 	}
 
-	UDlgContext* Context = NewObject<UDlgContext>(Participants[0], UDlgContext::StaticClass());
-	return Context->CanBeStarted(Dialogue, ParticipantBinding);
+	return UDlgContext::CanBeStarted(Dialogue, ParticipantBinding);
 }
 
 UDlgContext* UDlgManager::ResumeDialogue(